fixed: when user was deleted, it was still staying in unverified users db. now deleted from both db

+++ server/routes/api/users.js
@@ -1,6 +1,7 @@
 const express = require("express")
 const router = express.Router()
 const User = require("../../models/user")
+const UnverifiedUser = require("../../models/unverifiedUser")
 
 router.get("/", async (req, res) => {
     try {
@@ -8,7 +9,7 @@ router.get("/", async (req, res) => {
         res.send(users)
     } catch (error) {
         console.log(error)
-        res.status(500).json({message: error.message})
+        res.status(500).json({ message: error.message })
     }
 })
 
@@ -26,11 +27,18 @@ router.delete("/:name", async (req, res) => { //change to username
     const { name } = req.params
     const userName = name.replace("@", "")
     try {
-        const user = await User.findOneAndDelete({ name: userName })
-        res.status(201).json({ message: "Successfully deleted" });
+        const user = await User.findOne({ name: userName })
+        if (!user) {
+            return res.status(404).json({ message: "User not found" })
+        }
+
+        await User.findOneAndDelete({ name: userName })
+        await UnverifiedUser.findOneAndDelete({ email: user.email })
+
+        res.status(200).json({ message: "Successfully deleted" })
     } catch (error) {
-        console.log("error in api")
-        res.json({message: error})
+        console.error(error)
+        res.status(500).json({ message: "Server error" })
     }
 })
